Add psc-package definitions to build this package.
This is not strictly necessary for inclusion in psc-package, but it
might help resolve this library not building on travis:

It might work to install psc-package, instead of bower, and running
psc-package build on travis (it worked locally).

I'm also working on having remote-data included in the package-set. I
almost developed something very similar, so I prefer to use this
instead :-).
.